refactor(web): group settings into Workflow / Agents / Advanced sections
Extract the flat settings panel into a shared useSettingsForm hook plus
three category-scoped section components:

- Workflow: auto_review, require_review_approval, auto_confirm_single_tasks,
  review_strictness, planning_depth, serialize_merges, default_base_branch,
  worktree_base_ref_strategy.
- Agents: default_agent_backend, use_recommended_backend,
  default_model_claude, default_model_openai, additional_instructions.
- Advanced: theme, git_user_mode/name/email, auto_init_git_repo,
  auto_init_git_initial_commit, open_last_project_on_startup,
  attached_launcher, skip_attached_instructions_popup.

Setting keys, validation, and save behavior are unchanged — each field
still saves through the same apiClient.setSettings call. The ToggleRow
primitive is lifted out of the old panel so section components do not
depend on one another.

The unused settings/sections/*.tsx modules and the monolithic
settings-panel.tsx are removed; neither was imported by the live page.
